requirements.txt
Django>=4.0,<5.0
djangorestframework>=3.14.0
gunicorn>=20.1.0
